Implementación de servicios, y recibimiento de la API del front para envío de materias

+++ app/repositories/courses_dao.py
@@ -1,5 +1,5 @@
 from db.base_class import Base
-from sqlalchemy import Column, Integer, String, CheckConstraint, ForeignKey
+from sqlalchemy import Column, Integer, String, Float, ForeignKey
 from sqlalchemy.orm import relationship
 
 class CourseDAO(Base):
@@ -8,7 +8,7 @@ class CourseDAO(Base):
     id = Column(Integer, primary_key=True, unique=True, nullable=False)
     subject_id = Column(String(10), ForeignKey('subject.id'), nullable=False)
     teacher_id = Column(Integer, ForeignKey('teachers.id'), nullable=False)
-    difficulty = Column(Integer, nullable=False)
+    difficulty = Column(Float, nullable=False)
 
     subject = relationship("SubjectDAO", back_populates="courses")
     teacher = relationship("TeachersDAO", back_populates="courses")
